add: excetions on cliente services

+++ backend/trainee_devinhouse/src/main/java/br/com/devinhouse/trainee/services/ClienteServices.java
@@ -20,19 +20,19 @@ public class ClienteServices {
 	}
 	
 	// Validar quantidade de digitos do CPF
-	protected boolean validateCPF(Long cpf) {
-		boolean isValid = String.valueOf(cpf).length() == 11;
+	protected boolean validateCPF(String cpf) {
+		boolean isValid = cpf.length() == 11;
 		
 		return isValid;
 	}
 	
 	// Localizar cliente atraves do CPF
-	protected Cliente foundClientByCPF(Long cpf) {
+	protected Cliente foundClientByCPF(String cpf) {
 		Cliente foundClient = new Cliente();
 		List<Cliente> allClients = findAllClients();
 		
 		for(Cliente each : allClients) {
-			if(cpf.equals(each.getCPF())) {
+			if(cpf.equals(each.getCpf())) {
 				foundClient = each;
 			}
 		}
@@ -40,9 +40,28 @@ public class ClienteServices {
 		return foundClient;
 	}
 	
+	// Verifica se o CPF esta cadastrado
+	protected boolean cpfIsActive(String cpf) {
+		boolean status = false;
+		Cliente foundClient = foundClientByCPF(cpf);
+		
+		if (cpf.equals(foundClient.getCpf())) {
+			status = true;
+		}
+		
+		return status;
+	}
+	
 	// Cadastra cliente;
 	public Cliente create(Cliente obj) {
-		return clienteRepository.save(obj);
+		Cliente clientObj = new Cliente();
+		
+		if(validateCPF(obj.getCpf())) {
+			clienteRepository.save(obj);
+			clientObj = foundClientByCPF(obj.getCpf());
+			
+			return clientObj;
+		} throw new RuntimeException("Verifique a quantidade de digitos do CPF");
 	}
 	
 	// Busca todos os clientes;
@@ -51,32 +70,38 @@ public class ClienteServices {
 	}
 	
 	// Busca um cliente baseado no seu CPF
-	public Cliente getByCPFKey(Long cpf) {
-		return foundClientByCPF(cpf);
+	public Cliente getByCPFKey(String cpf) {
+		if (cpfIsActive(cpf)) {
+			return foundClientByCPF(cpf);
+		} throw new RuntimeException("CPF nao encontrado");
 	}
 		
 	// Atualiza os atributos de um cliente baseado no seu CPF
-	public Cliente updateClientByCPF(Long cpf, Cliente newObj) {
-		Cliente filteredClient = foundClientByCPF(cpf);
-		
-		if (newObj.getNome() != null) {
-			filteredClient.setNome(newObj.getNome());
-		}
-		if (newObj.getSobrenome() != null) {
-			filteredClient.setSobrenome(newObj.getSobrenome());
-		}
-		if (newObj.getBirthYear() != null) {
-			filteredClient.setBirthYear(newObj.getBirthYear());
-		}
-		
-		return filteredClient;
+	public Cliente updateClientByCPF(String cpf, Cliente newObj) {
+		if (cpfIsActive(cpf)) {
+			Cliente filteredClient = foundClientByCPF(cpf);
+			
+			if (newObj.getNome() != null) {
+				filteredClient.setNome(newObj.getNome());
+			}
+			if (newObj.getSobrenome() != null) {
+				filteredClient.setSobrenome(newObj.getSobrenome());
+			}
+			if (newObj.getBirthYear() != null) {
+				filteredClient.setBirthYear(newObj.getBirthYear());
+			}
+			
+			return filteredClient;
+		} throw new RuntimeException("Nao foi possivel atualizar o cliente, CPF nao encontrado");		
 	}
 
 	// Remove um cliente baseado no seu CPF
-	public List<Cliente> delete(Long cpf) {
-		Cliente filteredClient = foundClientByCPF(cpf);
-		clienteRepository.delete(filteredClient);
-		
-		return findAllClients();		
-	}	
+	public List<Cliente> delete(String cpf) {
+		if (cpfIsActive(cpf)) {
+			Cliente filteredClient = foundClientByCPF(cpf);
+			clienteRepository.delete(filteredClient);
+			
+			return findAllClients();		
+		} throw new RuntimeException("Nao foi possivel remover o cliente, CPF nao encontrado");
+	}
 }
